+++ app/service/model.py
@@ -1,8 +1,8 @@
-from app import db
+from app import db, ma
 from base_model import Base
 
 
-class Services(Base,db.Model):
+class Services(Base, db.Model):
     __tablename__ = 'services'
     __table_args__ = {'extend_existing': True}
     
@@ -13,3 +13,9 @@ class Services(Base,db.Model):
     duration = db.Column(db.Integer)
     user_id = db.Column(db.Integer, db.ForeignKey('user.id'))
 
+class Serviceschema(ma.Schema):
+    class Meta:
+        fields = ('id', 'style', 'description', 'cost', 'duration', 'user_id')
+        
+service_schema = Serviceschema()
+services_schemas = Serviceschema(many=True)
